types: move proto aliases from api package
