+++ abci/types/messages_test.go
@@ -9,7 +9,7 @@ import (
 	"github.com/cosmos/gogoproto/proto"
 	"github.com/stretchr/testify/assert"
 
-	cmtproto "github.com/cometbft/cometbft/api/cometbft/types"
+	cmtproto "github.com/cometbft/cometbft/types/proto"
 )
 
 func TestMarshalJSON(t *testing.T) {
