+++ evidence/reactor_test.go
@@ -15,7 +15,6 @@ import (
 
 	dbm "github.com/cometbft/cometbft-db"
 
-	cmtproto "github.com/cometbft/cometbft/api/cometbft/types"
 	cfg "github.com/cometbft/cometbft/config"
 	"github.com/cometbft/cometbft/crypto"
 	"github.com/cometbft/cometbft/crypto/tmhash"
@@ -26,6 +25,7 @@ import (
 	p2pmocks "github.com/cometbft/cometbft/p2p/mocks"
 	sm "github.com/cometbft/cometbft/state"
 	"github.com/cometbft/cometbft/types"
+	cmtproto "github.com/cometbft/cometbft/types/proto"
 )
 
 var (
