+++ privval/signer_requestHandler.go
@@ -4,11 +4,11 @@ import (
 	"fmt"
 
 	cryptoproto "github.com/cometbft/cometbft/api/cometbft/crypto/v1beta1"
-	cmtproto "github.com/cometbft/cometbft/api/cometbft/types"
 	"github.com/cometbft/cometbft/crypto"
 	cryptoenc "github.com/cometbft/cometbft/crypto/encoding"
 	pvproto "github.com/cometbft/cometbft/privval/proto"
 	"github.com/cometbft/cometbft/types"
+	cmtproto "github.com/cometbft/cometbft/types/proto"
 )
 
 func DefaultValidationRequestHandler(
