+++ types/canonical.go
@@ -3,7 +3,7 @@ package types
 import (
 	"time"
 
-	cmtproto "github.com/cometbft/cometbft/api/cometbft/types"
+	cmtproto "github.com/cometbft/cometbft/types/proto"
 	cmttime "github.com/cometbft/cometbft/types/time"
 )
