+++ types/canonical_test.go
@@ -4,21 +4,29 @@ import (
 	"reflect"
 	"testing"
 
-	cmtproto "github.com/cometbft/cometbft/api/cometbft/types"
 	"github.com/cometbft/cometbft/crypto/tmhash"
 	cmtrand "github.com/cometbft/cometbft/libs/rand"
+	cmtproto "github.com/cometbft/cometbft/types/proto"
 )
 
 func TestCanonicalizeBlockID(t *testing.T) {
 	randhash := cmtrand.Bytes(tmhash.Size)
-	block1 := cmtproto.BlockID{Hash: randhash,
-		PartSetHeader: cmtproto.PartSetHeader{Total: 5, Hash: randhash}}
-	block2 := cmtproto.BlockID{Hash: randhash,
-		PartSetHeader: cmtproto.PartSetHeader{Total: 10, Hash: randhash}}
-	cblock1 := cmtproto.CanonicalBlockID{Hash: randhash,
-		PartSetHeader: cmtproto.CanonicalPartSetHeader{Total: 5, Hash: randhash}}
-	cblock2 := cmtproto.CanonicalBlockID{Hash: randhash,
-		PartSetHeader: cmtproto.CanonicalPartSetHeader{Total: 10, Hash: randhash}}
+	block1 := cmtproto.BlockID{
+		Hash:          randhash,
+		PartSetHeader: cmtproto.PartSetHeader{Total: 5, Hash: randhash},
+	}
+	block2 := cmtproto.BlockID{
+		Hash:          randhash,
+		PartSetHeader: cmtproto.PartSetHeader{Total: 10, Hash: randhash},
+	}
+	cblock1 := cmtproto.CanonicalBlockID{
+		Hash:          randhash,
+		PartSetHeader: cmtproto.CanonicalPartSetHeader{Total: 5, Hash: randhash},
+	}
+	cblock2 := cmtproto.CanonicalBlockID{
+		Hash:          randhash,
+		PartSetHeader: cmtproto.CanonicalPartSetHeader{Total: 10, Hash: randhash},
+	}
 
 	tests := []struct {
 		name string
