+++ types/priv_validator.go
@@ -5,9 +5,9 @@ import (
 	"errors"
 	"fmt"
 
-	cmtproto "github.com/cometbft/cometbft/api/cometbft/types"
 	"github.com/cometbft/cometbft/crypto"
 	"github.com/cometbft/cometbft/crypto/ed25519"
+	cmtproto "github.com/cometbft/cometbft/types/proto"
 )
 
 // PrivValidator defines the functionality of a local CometBFT validator
