+++ types/validator.go
@@ -6,10 +6,10 @@ import (
 	"fmt"
 	"strings"
 
-	cmtproto "github.com/cometbft/cometbft/api/cometbft/types"
 	"github.com/cometbft/cometbft/crypto"
 	ce "github.com/cometbft/cometbft/crypto/encoding"
 	cmtrand "github.com/cometbft/cometbft/libs/rand"
+	cmtproto "github.com/cometbft/cometbft/types/proto"
 )
 
 // Volatile state for each Validator
